+++ mobile/ParamEditBool.qml
@@ -73,6 +73,8 @@ Item {
 
             Switch {
                 id: boolSwitch
+                focusPolicy: Qt.NoFocus
+
                 Layout.fillWidth: true
 
                 onCheckedChanged: {
@@ -89,6 +91,8 @@ Item {
                 Layout.fillWidth: true
                 Button {
                     id: nowButton
+                    focusPolicy: Qt.NoFocus
+
                     Layout.fillWidth: true
                     Layout.preferredWidth: 500
                     flat: true
@@ -101,6 +105,8 @@ Item {
 
                 Button {
                     id: defaultButton
+                    focusPolicy: Qt.NoFocus
+
                     Layout.fillWidth: true
                     Layout.preferredWidth: 500
                     flat: true
@@ -113,6 +119,8 @@ Item {
 
                 Button {
                     id: helpButton
+                    focusPolicy: Qt.NoFocus
+
                     Layout.fillWidth: true
                     Layout.preferredWidth: 500
                     flat: true