+++ mobile/ParamEditEnum.qml
@@ -77,6 +77,7 @@ Item {
 
             ComboBox {
                 id: enumBox
+                focusPolicy: Qt.NoFocus
                 Layout.fillWidth: true
 
                 background: Rectangle {
@@ -101,6 +102,8 @@ Item {
                 Layout.fillWidth: true
                 Button {
                     id: nowButton
+                    focusPolicy: Qt.NoFocus
+
                     Layout.fillWidth: true
                     Layout.preferredWidth: 500
                     flat: true
@@ -113,6 +116,8 @@ Item {
 
                 Button {
                     id: defaultButton
+                    focusPolicy: Qt.NoFocus
+
                     Layout.fillWidth: true
                     Layout.preferredWidth: 500
                     flat: true
@@ -125,6 +130,8 @@ Item {
 
                 Button {
                     id: helpButton
+                    focusPolicy: Qt.NoFocus
+
                     Layout.fillWidth: true
                     Layout.preferredWidth: 500
                     flat: true